[#WEB-717] Date/Time Range Query parameter preset (after login) should be configurable

+++ dcm4chee/dcm4chee-web/trunk/dcm4chee-web-war/src/main/java/org/dcm4chee/web/war/config/delegate/WebCfgDelegate.java
@@ -364,7 +364,7 @@ public class WebCfgDelegate extends BaseCfgDelegate {
             cal.set(Calendar.MINUTE, 0);
             cal.set(Calendar.SECOND, 0);
             cal.set(Calendar.MILLISECOND, 0);
-            long start = cal.getTimeInMillis() - offset * 86400000;
+            long start = cal.getTimeInMillis() - (long)offset * 86400000l;
             cal.set(Calendar.HOUR_OF_DAY, 23);
             cal.set(Calendar.MINUTE, 59);
             cal.set(Calendar.SECOND, 59);
